+++ CHANGES.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.8.0] - 2025-05-01
+
 ### Changed
 
 - (BREAKING) Prohibit `AmazingActivist::Base#new` calls. All activities must be
@@ -106,7 +109,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[Unreleased]: https://github.com/ixti/amazing-activist/compare/v0.7.0...main
+[Unreleased]: https://github.com/ixti/amazing-activist/compare/v0.8.0...main
+[0.8.0]: https://github.com/ixti/amazing-activist/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/ixti/amazing-activist/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/ixti/amazing-activist/compare/v0.5.2...v0.6.0
 [0.5.2]: https://github.com/ixti/amazing-activist/compare/v0.5.1...v0.5.2